모듈 파일로 분리
