+++ src/basic/pages/cart/events/handleRemoveCartItem.js
@@ -1,0 +1,13 @@
+import { cartItemStore } from '../store.js';
+
+export function handleRemoveCartItem(event) {
+  const target = event.target;
+
+  cartItemStore.setState((prevState) => {
+    const updatedCartItems = prevState.cartItems.filter(
+      (item) => item.id !== target.dataset.itemId
+    );
+
+    return { cartItems: updatedCartItems };
+  });
+}
